+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - Slightly improve performance of grapheme cluster segmentation.
 - Fixes grapheme cluster segmentation of multiple consecutive regional flags.
+- Add access to Age property of a codepoint (giving information about at which Unicode version a codepoint was introduced).
 - CMake install target also installs header and library (not just tools).
 
 ## 0.1.0 (2022-11-03)